Use optional Int for AutoTitleRec startYear

+++ src/main/scala/com/iscs/ratingbunny/domains/package.scala
@@ -26,7 +26,7 @@ package object domains:
   final case class AutoTitleRec(
       _id: String,
       primaryTitle: String,
-      startYear: Option[Double] = None,
+      startYear: Option[Int] = None,
       rating: Option[TitleRating] = None
   ) extends AutoRecBase
